requirements.txt
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
python-dotenv==0.10.5
six==1.14.0
visitor==0.1.3
websocket-client==0.57.0
websockets==8.1
Werkzeug==0.16.0
